requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
bidict==0.21.4
charset-normalizer==2.0.12
click==8.0.4
dnspython==2.2.0
eventlet==0.33.0
Flask==2.0.3
Flask-SocketIO==5.1.1
frozenlist==1.3.0
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
multidict==6.0.2
python-engineio==4.3.1
python-socketio==5.5.2
six==1.16.0
Werkzeug==2.0.3
yarl==1.7.2
